Move `to_batches` into `chia.util.misc`
