+++ tests/plot_sync/test_sync_simulated.py
@@ -28,8 +28,8 @@ from chia.server.ws_connection import WSChiaConnection
 from chia.simulator.block_tools import BlockTools
 from chia.simulator.time_out_assert import time_out_assert
 from chia.types.blockchain_format.sized_bytes import bytes32
-from chia.util.generator_tools import to_batches
 from chia.util.ints import int16, uint64
+from chia.util.misc import to_batches
 from tests.plot_sync.util import start_harvester_service
 
 log = logging.getLogger(__name__)
